Build a universal LuaJIT
LuaJIT does not like being built as a universal binary in a single
build. Hence, we build it once for each value in
`CMAKE_OSX_ARCHITECTURES` and then `lipo` them all up into a universal
binary when we're done.

This approach ends up requiring an extra build of LuaJIT (in the final
`BuildLuajit` call), but we currently need this to make sure that the
LuaJIT `install` target works properly in the directory that we've
`lipo`ed everything into.