Pass `CMAKE_OSX_ARCHITECTURES` correctly to sub-cmake calls
`CMAKE_OSX_ARCHITECTURES` contains a semi-colon, which causes problems
when naively passed to deps builds that use CMake.

We get around it by exploiting the `LIST_SEPARATOR` argument of
`ExternalProject_Add`.

This hack was borrowed from

    https://stackoverflow.com/a/45433229

An alternative fix to this, which will require a little more
refactoring, would be to use the `CMAKE_CACHE_ARGS` variable instead.
See

    https://www.scivision.dev/cmake-externalproject-list-arguments/