ci(release): mangle `gettext` more selectively
We actually only want to hide the shared libraries provided by gettext
to guarantee static linkage, so let's limit ourselves to deleting those.
We have to be careful to only delete the symlinks or else we risk
breaking gettext-provided binaries.

Also, let's get rid of `Mono.framework`. It contains headers and
libraries that shadow those provided by `gettext`, and we don't want
that.